# docs/openapi.json
{"openapi":"3.0.1","info":{"title":"Titanic API","description":"This is API provide multiple functionality endpoints over titanic dataset","contact":{"name":"Eli Bracha"},"version":"1.0"},"servers":[{"url":"/api/v1"}],"paths":{"/health":{"get":{"tags":["health"],"summary":"Get health check status","description":"Get health check status","operationId":"healthcheck","responses":{"200":{"description":"OK","content":{}}}}},"/passenger":{"get":{"tags":["passenger"],"summary":"Get passengers","description":"Get all passengers","operationId":"passenger-get-all","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/passenger.Response"}}}}},"500":{"description":"Internal Server Error","content":{}}}}},"/passenger/fare/histogram/percentile":{"get":{"tags":["passenger"],"summary":"Get fare histogram histogram","description":"Get histogram represention of number of passengers in each precentile","operationId":"passenger-fare-histogram","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/histogram.Histogram"}}}},"500":{"description":"Internal Server Error","content":{}}}}},"/passenger/{id}":{"get":{"tags":["passenger"],"summary":"Get passenger","description":"Get passenger by ID number","operationId":"passenger-get","parameters":[{"name":"id","in":"path","description":"Passenger ID","required":true,"schema":{"type":"integer"}},{"name":"attributes","in":"query","description":"Allowed: id, age, sex, name, survived, class, siblings-spouses, parents-children, ticket, fare, cabin, embarked","style":"form","explode":false,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/passenger.Response"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/errors.Error"}}}},"500":{"description":"Internal Server Error","content":{}}}}}},"components":{"schemas":{"errors.Error":{"type":"object","properties":{"code":{"type":"integer"},"message":{"type":"string"}}},"histogram.Entry":{"type":"object","properties":{"bin":{"type":"integer"},"count":{"type":"integer"}}},"histogram.Histogram":{"type":"object","properties":{"entries":{"type":"array","items":{"$ref":"#/components/schemas/histogram.Entry"}}}},"passenger.Response":{"type":"object","properties":{"age":{"type":"string"},"cabin":{"type":"string"},"class":{"type":"integer"},"embarked":{"type":"string"},"fare":{"type":"number"},"id":{"type":"integer"},"name":{"type":"string"},"parents-children":{"type":"integer"},"sex":{"type":"string"},"siblings-spouses":{"type":"integer"},"survived":{"type":"integer"},"ticket":{"type":"string"}}}}}}